Cast i2c frequency to uint32_t

+++ src/drv/touch/touch_driver_gt911.h
@@ -91,7 +91,7 @@ class TouchGt911 : public BaseTouch {
 
     void init(int w, int h)
     {
-        Wire.begin(TOUCH_SDA, TOUCH_SCL, I2C_TOUCH_FREQUENCY);
+        Wire.begin(TOUCH_SDA, TOUCH_SCL, (uint32_t)I2C_TOUCH_FREQUENCY);
         // delay(300); // already happens in touch.begin()
         touch_scan(Wire);
